backend/requirements.txt
fastapi
uvicorn[standard]
icalendar
cachetools
pytz
python-dotenv
